Fix assorted bugs across eco-api
1. Fix infinite recursion in Entities.getEntity() - the method called
   itself unconditionally, causing StackOverflowError. The registry
   iteration that follows already handles custom entity lookup.

2. Fix PriceItem.pay() removing more items than required - compared
   each stack against the total amount instead of the remaining amount
   needed, overcharging players when items spanned multiple slots.

3. Fix inverted logic in CommandBase.notifyOfflinePlayerRequired() -
   notifyFalse(!hasPlayedBefore, key) rejected players who have played
   and accepted players who haven't. Removed erroneous negation.

4. Fix double parseArguments() call in Items.doParse() - the result
   was checked for null but then parseArguments was called again
   instead of reusing the stored result, causing double mutation of
   the ItemMeta.

5. Fix resource leak in UpdateChecker.getVersion() - InputStream and
   Scanner were never closed. Wrapped in try-with-resources.

6. Delete unused Paste class - deprecated since 6.77.0, has zero
   dependents, and the default haste server has been offline for years.

7. Fix NumberUtils.format() int overflow - cast to (int) silently
   corrupted values exceeding Integer.MAX_VALUE. Changed to (long).

8. Fix gradient null array for single-character strings -
   getGradientColors() returned a null entry when step was 1,
   causing the literal text "null" to appear in output.

9. Fix missing space in EcoPlugin version error message - "You must
   be on at least" was concatenated directly with the version number.

+++ eco-api/src/main/java/com/willfp/eco/util/NumberUtils.java
@@ -200,7 +200,7 @@ public final class NumberUtils {
         DecimalFormat df = new DecimalFormat("0.00");
         String formatted = df.format(toFormat);
 
-        return formatted.endsWith("00") ? String.valueOf((int) toFormat) : formatted;
+        return formatted.endsWith("00") ? String.valueOf((long) toFormat) : formatted;
     }
 
     /**
